Newer UMD definitions for resources, updates to resource names
