+++ src/Core/Resources/twist/debug/js/twistdebug.js
