+++ src/Core/Resources/twist/manager/scss/twistmanager.scss
